refactor: message service -> interface with statistics message service
