+++ src/main/kotlin/dev/farneser/tasktracker/emailsender/dto/StatisticsDto.kt
@@ -2,6 +2,6 @@ package dev.farneser.tasktracker.emailsender.dto
 
 import jakarta.validation.constraints.Email
 
-data class StatisticDto(@Email val email: String) {
+data class StatisticsDto(@Email val email: String) {
     val columns: ArrayList<ColumnDto> = ArrayList()
 }
