+++ src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,0 +1,9 @@
+{
+  "properties": [
+    {
+      "name": "application.client.conifrmation.url",
+      "type": "java.lang.String",
+      "description": "The URL endpoint (with query) to which the client will be redirected after a successful authentication."
+    }
+  ]
+}
