+++ src/main/resources/application-dev.properties
@@ -14,3 +14,4 @@ spring.mail.password=password
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.starttls.required=true
+application.client.conifrmation.url=${CLIENT_CONFIRMATION_URL:http://localhost:3000/confirm?token=}
