+++ src/main/resources/application.properties
@@ -15,3 +15,4 @@ spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.mail.properties.mail.smtp.starttls.required=true
 logging.level.root=${LOG_LEVEL:INFO}
+application.client.conifrmation.url=${CLIENT_CONFIRMATION_URL:http://localhost:3000/confirm?token=}
